{{cookiecutter.project_name}}/pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "{{ cookiecutter.project_name }}"
description = "{{ cookiecutter.project_verbose_name }}"
version = "0.1.0"
readme = "README.md"
authors = ["{{ cookiecutter.project_domain }}"]

[tool.poetry.dependencies]
python = "^3.9"

django = "^3.2"
django-hurricane = "^0.8.8"
django-split-settings = "^1.0"
django-csp = "^3.7"
django-http-referrer-policy = "^1.1"
django-permissions-policy = "^4.1"
django-extensions = "^3.1.3"
django-storages = "^1.11.1"

psycopg2-binary = "<2.9"
python-decouple = "^3.4"
bcrypt = "^3.2"
structlog = "^21.1"


[tool.poetry.dev-dependencies]
django-debug-toolbar = "^3.2"
django-querycount = "^0.7"
django-coverage-plugin = "^2.0"

flake8-pytest-style = "^1.5"
flake8-django = "^1.1"
flake8-logging-format = "^0.6"

pytest = "^6.2"
pytest-django = "^4.4"
pytest-cov = "^2.12"
pytest-randomly = "^3.8"
pytest-deadfixtures = "^2.2"
pytest-testmon = "^1.1"
pytest-timeout = "^1.4"
django-test-migrations = "^1.1"
hypothesis = "^6.14"

mypy = "^0.910"

sphinx = "^4.0"
sphinx-autodoc-typehints = "^1.12"
tomlkit = "^0.7"
doc8 = "^0.8"

yamllint = "^1.26"
safety = "^1.10"
polint = "^0.4"
ipython = "^7.25"
brunette = "^0.2.0"
